requirements.txt
psycopg2~=2.9.1
pymongo~=3.12.0

grpcio
grpcio_tools
concurrent
pymongo
grpc_requests
urllib
json